+++ src/components/shared/StoryCard.tsx
@@ -6,53 +6,8 @@ import {
   CardTitle,
 } from "@/components/ui/card"
 import Link from "next/link";
-
-type StoryCardProps = {
-  storyID: number,
-  title: string,
-  authorName: string,
-  authorID: number,
-  tags: Tag[],
-  summary: string,
-  rating: string,
-  coverImg: string,
-  words: number,
-  views: number,
-  stars: number
-}
-
-type Tag = {
-  name: string,
-  type: string
-}
-
-function getRatingColor(rating: string): string {
-  switch (rating) {
-    case "E":
-      return "bg-green-500";
-    case "T":
-      return "bg-yellow-500";
-    case "M":
-      return "bg-red-500";
-  }
-  return "";
-}
-
-function getTagColor(tagType: string): string {
-  switch (tagType) {
-    case "fandom":
-      return "bg-purple-600";
-    case "warning":
-      return "bg-orange-700";
-    case "genre":
-      return "bg-blue-600";
-    case "character":
-      return "bg-green-600";
-    case "other":
-      return "bg-gray-600";
-  }
-  return "";
-}
+import { Star, StarHalf, Star as StarEmpty } from "lucide-react";
+import { getRatingColor, getTagColor } from "@/hooks/useStoryCard";
 
 function TagDisp(props: Tag) {
   const tagColor = getTagColor(props.type);
@@ -63,13 +18,29 @@ function TagDisp(props: Tag) {
   </Link>;
 }
 
+function StarRating({ stars }: { stars: number }) {
+  const totalStars = 5;
+  const roundedStars = Math.round(stars * 2) / 2; // nearest 0.5
+
+  return (
+    <div className="inline-flex align-text-top gap-0.5">
+      {Array.from({ length: totalStars }).map((_, i) => {
+        if (i + 1 <= roundedStars) {
+          return <Star key={i} size={16} className="text-orange-500" fill="currentColor" />;
+        } else if (i + 0.5 === roundedStars) {
+          return <StarHalf key={i} size={16} className="text-orange-500" fill="currentColor" />;
+        } else {
+          return <Star key={i} size={16} className="text-gray-300" />;
+        }
+      })}
+    </div>
+  );
+}
+
 export function StoryCard(props: StoryCardProps) {
   const ratingColor = getRatingColor(props.rating);
   const tagList = props.tags.map((tag,index)=><TagDisp key={index} name={tag.name} type={tag.type}/>);
   const summary = props.summary.split("\n").map((str,index)=><span key={index}>{str}<br/></span>);
-  const gapSize = 0.2; // i cbb to get the actual gap size
-  const totalGapSize = gapSize*(Math.ceil(props.stars)-1);
-  const starPercent = ((props.stars+totalGapSize)/(5+gapSize*4)*100)+"%";
   return (
     <Card className="w-full">
       <CardHeader>
@@ -95,14 +66,7 @@ export function StoryCard(props: StoryCardProps) {
       </CardContent>
       <CardFooter>
         <div className="rounded-sm bg-gray-100 w-full p-2 text-sm text-gray-800">
-          {props.words} words • {props.views} views •
-          <span 
-            className={"text-xl text-transparent m-1 bg-clip-text"}
-            style={{backgroundImage: `linear-gradient(90deg, orange ${starPercent}, gray ${starPercent})`}}
-          >
-            ★★★★★
-          </span>
-          {props.stars} stars
+          {props.words} words • {props.views} views • <StarRating stars={props.stars}/> {props.stars} stars
         </div>
       </CardFooter>
     </Card>
