+++ src/types/story.ts
@@ -1,0 +1,18 @@
+type StoryCardProps = {
+  storyID: number,
+  title: string,
+  authorName: string,
+  authorID: number,
+  tags: Tag[],
+  summary: string,
+  rating: string,
+  coverImg: string,
+  words: number,
+  views: number,
+  stars: number
+}
+
+type Tag = {
+  name: string,
+  type: string
+}
